# sdks/openapi.json
{
  "components": {
    "schemas": {
      "AccountMe": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          }
        },
        "required": [
          "account_id",
          "email",
          "email_verified"
        ],
        "title": "AccountMe",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Modelo padrão para respostas de erro.",
        "properties": {
          "detail": {
            "description": "Descrição do erro",
            "title": "Detail",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Código interno do erro",
            "title": "Error Code"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Timestamp do erro",
            "title": "Timestamp"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in_minutes": {
            "title": "Expires In Minutes",
            "type": "integer"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in_minutes"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "SignupRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Mínimo 10 caracteres, com letras e números",
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "SignupResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "email_verified": {
            "default": false,
            "title": "Email Verified",
            "type": "boolean"
          }
        },
        "required": [
          "account_id",
          "email"
        ],
        "title": "SignupResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyEmailRequest": {
        "properties": {
          "token": {
            "minLength": 8,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "VerifyEmailRequest",
        "type": "object"
      },
      "VerifyEmailResponse": {
        "properties": {
          "email_verified": {
            "default": true,
            "title": "Email Verified",
            "type": "boolean"
          }
        },
        "title": "VerifyEmailResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key gerada no portal self-service, enviada como 'Authorization: Bearer <key>'. Requerida pelos endpoints de dados e busca semântica.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionAuth": {
        "description": "JWT de sessão do portal (login self-service). Não usar para chamadas de API de dados/IA — use ApiKeyAuth.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API pública que expõe toda a Base Nacional Comum Curricular (BNCC) do Brasil: dados determinísticos das três etapas, acesso self-service por API keys, documentação automática e busca semântica com IA (não-oficial).",
    "title": "BNCC API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "Autentica com e-mail/senha e retorna um JWT de sessão do portal (também definido como cookie httponly `session`). Falha de credencial ou e-mail não verificado retornam a mesma mensagem (anti-enumeração — Princípio V / FR-023).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "padrao": {
                  "summary": "Login padrão",
                  "value": {
                    "email": "dev@example.com",
                    "password": "SenhaForte123"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Token de sessão (JWT); também é definido como cookie httponly."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Credenciais inválidas ou e-mail não verificado."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Muitas tentativas de login deste IP."
          }
        },
        "summary": "Autenticar e iniciar sessão do portal",
        "tags": [
          "Autenticação"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "description": "Retorna os dados da conta associada à sessão do portal atual (header `Authorization: Bearer <jwt>` ou cookie `session`).",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountMe"
                }
              }
            },
            "description": "Dados básicos da conta do desenvolvedor autenticado."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Sessão ausente, inválida ou expirada."
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Dados da conta autenticada",
        "tags": [
          "Autenticação"
        ]
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "description": "Cria uma conta de desenvolvedor do portal e dispara um e-mail de verificação de uso único. A conta permanece `email_verified=false` até a confirmação (necessária para gerar API keys — FR-007). Não usa API key, apenas e-mail/senha.",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "padrao": {
                  "summary": "Cadastro padrão",
                  "value": {
                    "email": "dev@example.com",
                    "password": "SenhaForte123"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            },
            "description": "Conta criada (ainda não verificada); e-mail de verificação disparado."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Payload inválido (e-mail malformado ou senha fora da política)."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Não foi possível concluir o cadastro (e-mail já em uso)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Muitas tentativas de cadastro deste IP."
          }
        },
        "summary": "Criar conta de desenvolvedor",
        "tags": [
          "Autenticação"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "description": "Consome o token de verificação de uso único enviado por e-mail no cadastro e marca a conta como verificada. Necessário antes de gerar API keys (FR-007).",
        "operationId": "verify_email_api_v1_auth_verify_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "padrao": {
                  "summary": "Token recebido por e-mail",
                  "value": {
                    "token": "a1b2c3d4e5f6"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponse"
                }
              }
            },
            "description": "Confirmação de que o e-mail da conta foi verificado."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de verificação inválido."
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de verificação expirado ou já utilizado."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Muitas tentativas deste IP."
          }
        },
        "summary": "Confirmar e-mail via token de verificação",
        "tags": [
          "Autenticação"
        ]
      }
    },
...
  }
}